rft: 库存Item拆分
